+++ ext_tables.php
@@ -1,18 +1,7 @@
 <?php
 defined('TYPO3') || die();
 
-(static function () {
-    if ((new TYPO3\CMS\Core\Information\Typo3Version())->getMajorVersion() < 11) {
-        $iconRegistry = TYPO3\CMS\Core\Utility\GeneralUtility::makeInstance(
-            TYPO3\CMS\Core\Imaging\IconRegistry::class
-        );
-        $iconRegistry->registerIcon(
-            'form-country-select',
-            TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
-            ['source' => 'EXT:' . Brotkrueml\FormCountrySelect\Extension::KEY . '/Resources/Public/Icons/form-country-select.svg']
-        );
-    }
+use Brotkrueml\FormCountrySelect\Extension;
 
-    $GLOBALS['TYPO3_CONF_VARS']['SYS']['locallangXMLOverride']['EXT:form/Resources/Private/Language/Database.xlf'][]
-        = 'EXT:' . Brotkrueml\FormCountrySelect\Extension::KEY . '/Resources/Private/Language/Backend.xlf';
-})();
+$GLOBALS['TYPO3_CONF_VARS']['SYS']['locallangXMLOverride']['EXT:form/Resources/Private/Language/Database.xlf'][]
+    = 'EXT:' . Extension::KEY . '/Resources/Private/Language/Backend.xlf';